Releases 0.2.1 Fix sdist problem

+++ src/c_silkv3/coder.cpp
@@ -84,7 +84,7 @@ static PyObject *encode_silk(PyObject *self, PyObject *args) {
       break;
 
     counter = (frameSizeReadFromFile_ms * API_fs_Hz) / 1000;
-    if (counter > psReadEnd - psRead) {
+    if ((int)counter > psReadEnd - psRead) {
       memset(in, 0x00, sizeof(in));
 
       size_t realrd = (psReadEnd - psRead);
